Fix SR300 Max Z + Type casting
Fixed bug for the Maximum Z of SR300
- Should be 1.5m, was truncated to 1m
- Found with -Wdouble-promotion -Wfloat-conversion

Replaced old C-style casting to safer C++-style casting.

Fix camera test logic
- Use boolean instead of logic ANDing doubles

+++ realsense_camera/include/realsense_camera/base_nodelet.h
@@ -112,7 +112,7 @@ namespace realsense_camera
     std::string base_frame_id_;
     std::string depth_frame_id_;
     std::string color_frame_id_;
-    int max_z_ = -1;
+    float max_z_ = -1.0f;
     bool enable_pointcloud_;
     bool enable_tf_;
     bool duplicate_depth_color_;
